ddr: Change polling_needed to u32 to avoid SRAM alignment issues
In BL2, which runs from SRAM, if the `polling_needed` variabile
is kept as u8, with some compilers it might trigger a synchronous
exception when a 1-byte `strb` operation is issued on its address,
part of the `.data` section.

Make the `polling_needed` variable an u32 to avoid such issues
when running from SRAM.

Issue: ALB-10843

+++ drivers/nxp/s32/ddr/ddr_utils.c
@@ -46,7 +46,7 @@ static uint32_t adjust_ddrc_config(void);
 static bool is_lpddr4(void);
 
 #if (ERRATA_S32_050543 == 1)
-uint8_t polling_needed = 2;
+uint32_t polling_needed = 2;
 #endif
 
 static struct space_timing_params tr_res = {
