+++ plat/nxp/s32/include/s32_ddr_errata_funcs.h
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022 NXP
+ * Copyright 2022-2023 NXP
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 #ifndef S32_DDR_ERRATA_FUNCS_H
 #define S32_DDR_ERRATA_FUNCS_H
 
-void ddr_errata_update_flag(uint8_t flag);
+void ddr_errata_update_flag(uint32_t flag);
 
 #endif
